Bug: integer overflow for large files
Updates the gRPC protocol to allow downloading files that are over 2GB in size.
The previous version of the protocol was (naively) using int32 for offsets. So anything beyond the 2GB mark would produce an integer overflow server-side.

 - Rename deprecated unit32 fields, they will be ignored by current and future agent in favor of new int64
 - Add a field so that an agent can inform the server it is recent and supports files/ranges over 2GB
 - Block request whose range is over the 2GB mark IFF the target agent is old and does not support it
 - Migrate from int to long where appropriate

Old behavior:
 1) Files smaller than 2GB: no problem
 2) Range below 2GB for larger than 2GB file: no problem
 3) Range above 2GB for larger than 2GB files: error 500, integer overflow

New behavior
 1) Unchanged
 2) Unchanged
 3) (against old agent) Unchanged (but better error message)
 3) (against new agent) successful

+++ genie-agent/src/main/java/com/netflix/genie/agent/execution/services/impl/grpc/GRpcAgentFileStreamServiceImpl.java
@@ -255,8 +255,8 @@ public class GRpcAgentFileStreamServiceImpl implements AgentFileStreamService {
     private synchronized void handleFileRequest(
         final String streamId,
         final String relativePath,
-        final int startOffset,
-        final int endOffset
+        final long startOffset,
+        final long endOffset
     ) {
         log.info(
             "Server is requesting file {} (range: [{}, {}), streamId: {})",
@@ -350,19 +350,19 @@ public class GRpcAgentFileStreamServiceImpl implements AgentFileStreamService {
         private final GRpcAgentFileStreamServiceImpl gRpcAgentFileStreamService;
         private final String streamId;
         private final Path absolutePath;
-        private final int startOffset;
-        private final int endOffset;
+        private final long startOffset;
+        private final long endOffset;
         private final StreamObserver<AgentFileMessage> outboundStreamObserver;
         private final ByteBuffer readBuffer;
         private final AtomicBoolean completed = new AtomicBoolean();
-        private int watermark;
+        private long watermark;
 
         FileTransfer(
             final GRpcAgentFileStreamServiceImpl gRpcAgentFileStreamService,
             final String streamId,
             final Path absolutePath,
-            final int startOffset,
-            final int endOffset,
+            final long startOffset,
+            final long endOffset,
             final long maxChunkSize
         ) {
             this.gRpcAgentFileStreamService = gRpcAgentFileStreamService;
